ProjectFinal/blog/requirements.txt
asgiref==3.5.2
cycler==0.11.0
distlib==0.3.6
Django==4.0.6
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
ent==0.3.3
mysql==0.0.3
mysql-connector-python==8.0.30
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.4
py-mysql==1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.7
